SELF_JOIN
ABOUT THE TOPIC :
In MS-SQL, a SELF JOIN is when a table joins with itself to compare rows within the same table. It uses table aliases to differentiate instances and is often used for hierarchical data (e.g., employees and managers). It can be an INNER JOIN or LEFT JOIN based on requirements.

YouTube Link :  https://youtu.be/HXZjO3Z-_kU