SQL_ANY_ALL
ABOUT THE TOPIC :

In MS-SQL, `ANY` and `ALL` are used with comparison operators in subqueries to filter results based on a set of values. The `ANY` operator returns `TRUE` if at least one value in the subquery satisfies the condition, while `ALL` requires that all values meet the condition. These operators help compare a column's value against multiple results in a subquery efficiently.

YouTube Link : https://youtu.be/8vEEG2cKu88