SELECT INTO
ABOUT THE TOPIC :
In MS-SQL, the `SELECT INTO` statement is used to create a new table and insert the results of a `SELECT` query into it. This is useful for making backup copies of data or transforming data into a new structure without manually defining the table. The new table inherits column names and data types from the source query but does not include indexes or constraints.

YouTube Link : https://youtu.be/19uXYs9wR7U